minor type update | Need to look into kd_tree better in the future

+++ src/clusterxx/data_structures/kd_tree/kd_tree.hpp
@@ -21,7 +21,7 @@ class kd_tree {
         arma::mat __extra_points;
         std::vector<int> __extra_points_inds;
         size_t __ind;
-        int __feature_size;
+        size_t __feature_size;
 
         kd_node(const arma::vec &point, const size_t ind)
             : __point(point), __feature_size(__point.n_cols), __ind(ind) {}
